requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
av==16.0.1
beautifulsoup4==4.14.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
cymem==2.0.11
deep-translator==1.11.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
fastapi==0.120.4
faster-whisper==1.2.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.3
nltk==3.9.1
numpy==2.2.6
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
preshed==3.0.10
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
rich==14.2.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.49.3
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
torch
torchvision
torchaudio
--extra-index-url https://download.pytorch.org/whl/cpu
tqdm==4.67.1
transformers==4.54.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
v==1
wasabi==1.1.3
weasel==0.4.1
wrapt==2.0.0
